Add rules to plaintext CID
A few additional rules should make this even more robust to analysis.

The odd number foils this attack: I connect to a server and conduct a bunch of migrations to get lots of new connection IDs. If the divisor is an even number, the last bit of the routing mask will *always* be the same. This dramatically reduces the number of combinations I need to consider when deciphering the routing mask, and right away divides the server pool into two detectable halves.