No understandy = no worky
THe addition to the first paragraph here seems necessary.  The basic
idea is that if a QUIC version violates the (reasonable) assumptions
this section makes, then it just doesn't work if you try to apply this
design.  You will need to update your load balancer and probably update
the CID allocation/negotiation/configuration you are using.

The second paragraph I'm less sure about.  I think that the intent was
to say that if you follow the advice here you still get a functioning
system even if the load balancer gets it wrong.  So I said that.  I
think that I agree, provided that the tenant (new term?) chooses a good
connection ID for any new ones.  So that's not just Section 3.1 you need
to refer to, but the whole document.